Added VC++6 projects and lots of porting changes
git-svn-id: http://yate.null.ro/svn/yate/trunk@272 acf43c95-373e-0410-b603-e72c3f656dc1

+++ contrib/ysip/engine.cpp
@@ -271,7 +271,7 @@ void SIPEngine::processEvent(SIPEvent *event)
     delete event;
 }
 
-unsigned long long SIPEngine::getTimer(char which, bool reliable) const
+u_int64_t SIPEngine::getTimer(char which, bool reliable) const
 {
     switch (which) {
 	case '1':
